requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
mysql-connector==2.2.9
pypiwin32==223
pywin32==224
requests==2.22.0
urllib3==1.25.3
